Relationship between user and tasks

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { Task } from './task.entity';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 const TAG = 'tasks.service'
 @Injectable()
 export class TasksService {
@@ -18,15 +19,15 @@ export class TasksService {
         let res = await this.taskRepository.find()
         return res
     }
-    async getTaskById(id:number):Promise<Task>{
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(id:number,user:User):Promise<Task>{
+        const found = await this.taskRepository.findOne({where:{id,userId:user.id}})
         if(!found){
             throw new NotFoundException(`Task with id ${id} not found`)
         }
         return found
     }
-    async getTasks(filterDto:GetTasksFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto)
+    async getTasks(filterDto:GetTasksFilterDto,user:User):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto,user)
     }
     // getTasksWithFilters(filter:GetTasksFilterDto):Task[]{
     //     let res = this.getAllTasks()
@@ -48,13 +49,16 @@ export class TasksService {
     //     }
     //     return res
     // }
-    async createTask(createTaskDto:CreateTaskDto):Promise<Task>{
-        const task = await this.taskRepository.createTask(createTaskDto)
+    async createTask(
+        createTaskDto:CreateTaskDto,
+        user:User
+        ):Promise<Task>{
+        const task = await this.taskRepository.createTask(createTaskDto,user)
         return task
     }
-    async deleteTaskById(id:number):Promise<void>{
+    async deleteTaskById(id:number,user:User):Promise<void>{
         // let res = await this.getTaskById(id)
-        let res = await this.taskRepository.delete({id})
+        let res = await this.taskRepository.delete({id,userId:user.id})
         if(!res.affected){
             throw new NotFoundException(`Task with id ${id} not found`)
         }
@@ -70,9 +74,9 @@ export class TasksService {
     //     }
     //     return res
     // }
-    async updateTaskStatus(id:number,status:TaskStatus):Promise<Task>{
-        // let res2 = this.taskRepository.update({id:id},{status})
-        let res = await this.getTaskById(id)
+    async updateTaskStatus(id:number,status:TaskStatus,user:User):Promise<Task>{
+        let res2 = this.taskRepository.update({id:id},{status})
+        let res = await this.getTaskById(id,user)
         res.status = status
         await res.save()
         return res
